Prevent enter key submitting the form in windows
When using a RTE such as Redactor, the window has a tendency to submit the form when hitting enter. This breaks the ability to add line breaks in the editor.

This might actually be Redactor-specific as it defaults to an inline editor, while TinyMCE and others default to an iframe and the enter might not trigger in the form.